requirements.txt
aiogram
asyncio
http3
psycopg2-binary
fastapi[standard]
uvicorn
redis